+++ ChangeLog.md
@@ -1,7 +1,12 @@
 # Change Log - oav
 
-## 02/09/2023 3.2.5
+## 02/03/2023 3.2.5
 
+- Lock @apidevtools/swagger-parser@10.0.3 as 10.1.0 brings conflict of ajv package version
+- generate-api-scenario
+  - Support `--scope` option in static generator
+- API Scenario
+  - Skip create and delete resource group if specify resource group name
 - bug fix about incorrect payload file path in html report
 
 ## 01/30/2023 3.2.4